+++ leptos-fluent-macros/tests/ui/leptos_fluent/stable/fail/invalid_default_language.rs
@@ -1,0 +1,31 @@
+use fluent_templates::static_loader;
+use leptos::prelude::*;
+use leptos_fluent_macros::leptos_fluent;
+
+static_loader! {
+    pub static TRANSLATIONS = {
+        locales: "../../../../examples/csr-minimal/locales",
+        fallback_language: "en",
+    };
+}
+
+#[component]
+fn I18n(children: Children) -> impl IntoView {
+    leptos_fluent! {
+        children: children(),
+        translations: [TRANSLATIONS],
+        locales: "../../../../examples/csr-minimal/locales",
+        default_language: "foobarbaz",
+    }
+}
+
+#[component]
+pub fn App() -> impl IntoView {
+    view! {
+        <I18n>
+            <p>+</p>
+        </I18n>
+    }
+}
+
+fn main() {}